move handleOnClick from component to container
